[WIP] fx_box spec
